+++ e2e/commands/import.e2e.js
@@ -1243,7 +1243,8 @@ describe('bit import', function () {
       helper.reInitLocalScope();
       helper.addRemoteScope();
       helper.importComponent('utils/is-type@0.0.1');
-      helper.importComponent('utils/is-type@0.0.2');
+      // @todo: the --force flag here is a hack to work around a bug which shows the component of 0.0.1 as modified
+      helper.importComponent('utils/is-type@0.0.2 --force');
     });
     it('should imported v2 successfully and print the result from the latest version', () => {
       const appJsFixture = "const isType = require('./components/utils/is-type'); console.log(isType());";
@@ -1518,6 +1519,54 @@ describe('bit import', function () {
     it('should not create any link file', () => {
       localConsumerFiles.forEach((fileName) => {
         expect(fileName.includes('index.js')).to.be.false;
+      });
+    });
+  });
+
+  describe('import a component when the local version is modified', () => {
+    before(() => {
+      helper.setNewLocalAndRemoteScopes();
+      helper.createComponentBarFoo();
+      helper.addComponentBarFoo();
+      helper.commitAllComponents();
+      helper.exportAllComponents();
+      helper.reInitLocalScope();
+      helper.addRemoteScope();
+      helper.importComponent('bar/foo');
+      const barFooFixtureV2 = "module.exports = function foo() { return 'got foo v2'; };";
+      helper.createComponent(path.join('components', 'bar', 'foo', 'bar'), 'foo.js', barFooFixtureV2);
+
+      const appJsFixture = "const barFoo = require('./components/bar/foo'); console.log(barFoo());";
+      fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
+    });
+    describe('without --force flag', () => {
+      let output;
+      before(() => {
+        try {
+          helper.importComponent('bar/foo');
+        } catch (err) {
+          output = err.toString();
+        }
+      });
+      it('should display a warning saying it was unable to import', () => {
+        expect(output).to.have.string('unable to import');
+      });
+      it('should not override the local changes', () => {
+        const result = helper.runCmd('node app.js');
+        expect(result.trim()).to.equal('got foo v2');
+      });
+    });
+    describe('with --force flag', () => {
+      let output;
+      before(() => {
+        output = helper.importComponent('bar/foo --force');
+      });
+      it('should display a successful message', () => {
+        expect(output).to.have.string('successfully imported');
+      });
+      it('should override the local changes', () => {
+        const result = helper.runCmd('node app.js');
+        expect(result.trim()).to.equal('got foo');
       });
     });
   });